Made DirectionSteps more modular
